+++ modules/swagger-codegen/src/test/java/io/swagger/codegen/v3/config/CodegenConfiguratorTest.java
@@ -1,4 +1,4 @@
-package io.swagger.codegen.config;
+package io.swagger.codegen.v3.config;
 
 import io.swagger.codegen.v3.ClientOptInput;
 import io.swagger.codegen.v3.CodegenConfig;
